Implement a working version of the 'Plectrum' macro
This is not a complete implementation. Currently, the 'values' are
exactly the same as the tokens for variants i.e. in CamelCase which is
a convention in Rust. Ideally, we want the user to be able to specify
how the enum variant token is to be case-converted to get the value
e.g. lowecase, uppercase, snake_case etc. This will be implemented in
later commits.

A lot more can be improved from this point onwards.