Refactor code to delegate code generation to another function
Based on what I've observed in other crates that implement proc
macros, this seems to be a good practice. The function that's marked
as 'proc_macro_derive' parses the input token stream into AST and
delegates the actual code generation to another function. It seems
like this is done to separate the use of 'proc_macro' and
'proc_macro2' crates. The actual proc macro function deals with
'proc_macro' objects whereas the inner function deals with
'proc_macro2' objects.