Make the proc macro panic if the input type is not valid
The 'Plectrum' proc macro only works for non data-bearing enums. If
it's used on a struct or an enum with data-bearing variants, then the
code will panic.

Not sure if panicking is the best way to handle errors. But at least
the user experience in terms of compiler errors is decent
e.g. compiler error mentions the message, "The 'Plectrum' derive macro
only works for non data-bearing enums" in the help section.