Implement a default trait method in Enum trait to return an id
This method returns the 'id' corresponding to the enum variant given
the mapping as an argument. This API is much more intuitive as
compared to the previous 'Mapping.get_id' method.

The 'get_id' method of the Mapping struct is marked as
deprecated since version 0.2.0 (i.e. the next version; Even though the
version is not bumped yet, this works because rustc currently doesn't
interpret it)